Added a ChartGroup class to contain the Chart objects. All objects now load their settings from a configuration file.
Added code to handle data paths, job name, and job path.

The Filtered trace was moved to the second graph of the second chart as
the config file is only set up for two traces per chart. The existing
data generation will be replaced by inspection simulation...so it is just
used as a quick test for now.

Signed-off-by: Mike Schoonover <mike@skoonie.com>