Adjusted 3d map height to 180 which seems to be good for the 1920x1080 screen.
Fixed bug in configuration file where width was listed twice instead of
width and height.

Adjusted 3d display settings in Map3DGraph and MainController for best
results. Rotated the graph 180 degrees so position 0 is at the left.

Changed height of simulated signal on 3D graph from 25 to 15 to fit on
the graph vertically.

Signed-off-by: Mike Schoonover <mike@skoonie.com>