The 3D map parameters are now loaded from the config file. Currently MainController copies them to the control panel -- this needs to be moved to whatever code will open the control panel in response to clicking of the Calibrate button for that graph.
Option added to set the graph width/xPos settings in the config file to
force the grid to be automatically sized to fit the graph and to be
aligned to the left edge.

Changed Map3DGraph so it now only handles a single Map3D instead of
multiple maps. This was done to simplify coding and it would be difficult
to manage multiple maps in a single graph and probably wouldn't make sense.

Signed-off-by: Mike Schoonover <mike@skoonie.com>