Created a peak capture classes for arrays similar to the PeakBuffer class family for ints.
Added code to both to flag when the peak has not been updated so the
calling function will know the data is not valid.

Changed the original PeakBuffer family to be hard coded for ints instead
of using Objects, which was originally done to make the class more Generic
without using Generics as they only work with objects. Switching to ints
makes the code clearer and probably faster.

Renamed all classes for clarity and brevity. PeakBuffer class deleted as
it is no longer needed.

Signed-off-by: Mike Schoonover <mike@skoonie.com>