Added a wait between packet requests in initAfterConnect()
This was done because the master PIC was receiving the request all last
AD values command while he was still in the middle of doing stuff for the
request all status packet command. When this happens, the PIC just throws
away the request all last AD values command without ever handling it.

Signed-off-by: Hunter Schoonover <hunter@skoonie.com>