Java host now expects and handles Rbt and PIC run data pkt counts
The counts are checked for errors and then stored. If there are errors
found, they are recorded.

In "00-Test 1 Main Configuration.ini" buffer locations to 2 so that the
counts are not read as run data.

Simulation code now also sends the packet counts just like the Rabbit
would.

Finished Asana subtask ::
	Add Count Byte to Run Data Packet -> Add code to Java Host

Signed-off-by: Hunter Schoonover <hunter@skoonie.com>