Moving the mouse along a Trace in TraceGraph now changes ZoomGraph
When the mouse is moved, the greatest peak within +-5 of the x position is
now displayed. Currently, the y position of the mouse is not taken into
consideration. The code to do so is partially written, but not yet usable.
To allow the user to display different traces at the same x position and
view smaller peaks in an area, this must be finished out.

Signed-off-by: Hunter Schoonover <hunter@skoonie.com>