Fixed repeating snapshot bug when channel is turned off
lastRequestedPeak was being reset to 0 while lastRequestedPeakX was not
being reset at all. When the channel turned off, the peak returned to be
compared to lastRequestedPeak is 0, which is equal to and not greater than.
The lastRequestedPeakX was not being changed.

Problem was solved by resetting lastRequestedPeak to -1 instead of 0

Signed-off-by: Hunter Schoonover <hunter@skoonie.com>