Overall: Hardware should be flagging threshold violations as data is read
The flagging has not been tested AT ALL. The program works and allows for
the simulation mode to be ran in scan and inspect mode. Thresholds can
be adjusted up and down.

Threshold class now stores all of its information that may be read by other
classes and threads in an instance of ThresholdInfo. After the
ThresholdInfo object has been set up, he adds a reference to a list in
SharedSettings.

SharedSettings now contains a list of ThresholdInfos and has helper
functions to get a specific ThresholdInfo or get all for a particular graph

Device and all of its subclasses are now passed a reference to the
SharedSettings instance used throughout the entire program which is now
passed on to all Channels created for a Device.

Channel now makes use of SharedSettings by retrieving all of the
ThresholdInfos pertaining to thresholds on the same graph as the channel.
The data in the ThresholdInfos is used to determine if there are any
threshold violations.

Each peak now has his own threshold violation flag. In the Hardware/
Controller side, this is a simple boolean. When it is put into the Model,
I added DataTransferIntBuffer::storeThresholdAtInsertionPoint() to make use
of the flags buffer already in that class. View does not yet make use of
the threshold flag if it is set.

Signed-off-by: Hunter Schoonover <hunter@skoonie.com>