Works with timer driven tracking mode with ControlDevice an interface!!
The Multi_IO_A_Transverse class should also be able to be used as the
control device by simply setting timer driven tracking stuff in ini file to
false and by specifying the name of the control device.

However, this cannot ACTUALLY be done at this stage because there is no
simulation or real device code in place for a device to send inspect
packets back.

Signed-off-by: Hunter Schoonover <hunter@skoonie.com>