Ready to advance flag now reset properly and timer driven tracking works
Code used before that was reading in values from an external device and
using that to determine if the hardware was ready to advance insertion
points was not actually working properly because the ready to advance flag
was never being set back to false. It only worked because we were running
in simulation mode, which is so fast that the timing just happened to work.

Signed-off-by: Hunter Schoonover <hunter@skoonie.com>