Fixed SNAP_VALS_NO_MATCH_TRACES!!!
This check in is messy, has a lot of test code that needs to be removed.
Solution also needs to be cleaned up.

Solution is in MainController::displayDataFromDevices. Incrementing put
pointers is now done AFTER all of the channels have a chance to put data
into the current put pointer indexes.